Fix int32 offset overflow in kpool gather kernel

+++ python/sglang/srt/layers/attention/dsa/kpool_fp8_index.py
@@ -81,7 +81,8 @@ def _gather_index_k_scale_prefix_into_kernel(
     token_id = tl.program_id(0)
     page_idx = token_id // PAGE_SIZE
     token_offset_in_page = token_id % PAGE_SIZE
-    page = tl.load(page_indices_ptr + page_idx)
+    # Use int64 because page offsets can overflow int32 for large index buffers.
+    page = tl.load(page_indices_ptr + page_idx).to(tl.int64)
 
     offs = tl.arange(0, BLOCK_D)
     mask = offs < HEAD_DIM
